refactor: transform commerce
