+++ src/modules/commerce/module.ts
@@ -1,79 +1,7 @@
 import { FakerError } from '../../errors/faker-error';
 import { ModuleBase } from '../../internal/module-base';
-import { calculateUPCCheckDigit } from './_upc-check-digit';
-
-// Source for official prefixes: https://www.isbn-international.org/range_file_generation
-const ISBN_LENGTH_RULES: Record<
-  string,
-  Array<[rangeMaximum: number, length: number]>
-> = {
-  '0': [
-    [1999999, 2],
-    [2279999, 3],
-    [2289999, 4],
-    [3689999, 3],
-    [3699999, 4],
-    [6389999, 3],
-    [6397999, 4],
-    [6399999, 7],
-    [6449999, 3],
-    [6459999, 7],
-    [6479999, 3],
-    [6489999, 7],
-    [6549999, 3],
-    [6559999, 4],
-    [6999999, 3],
-    [8499999, 4],
-    [8999999, 5],
-    [9499999, 6],
-    [9999999, 7],
-  ],
-  '1': [
-    [99999, 3],
-    [299999, 2],
-    [349999, 3],
-    [399999, 4],
-    [499999, 3],
-    [699999, 2],
-    [999999, 4],
-    [3979999, 3],
-    [5499999, 4],
-    [6499999, 5],
-    [6799999, 4],
-    [6859999, 5],
-    [7139999, 4],
-    [7169999, 3],
-    [7319999, 4],
-    [7399999, 7],
-    [7749999, 5],
-    [7753999, 7],
-    [7763999, 5],
-    [7764999, 7],
-    [7769999, 5],
-    [7782999, 7],
-    [7899999, 5],
-    [7999999, 4],
-    [8004999, 5],
-    [8049999, 5],
-    [8379999, 5],
-    [8384999, 7],
-    [8671999, 5],
-    [8675999, 4],
-    [8697999, 5],
-    [9159999, 6],
-    [9165059, 7],
-    [9168699, 6],
-    [9169079, 7],
-    [9195999, 6],
-    [9196549, 7],
-    [9729999, 6],
-    [9877999, 4],
-    [9911499, 6],
-    [9911999, 7],
-    [9989899, 6],
-    [9999999, 7],
-  ],
-};
+import { ISBN_LENGTH_RULES } from './isbn';
+import { calculateUPCCheckDigit } from './upc';
 
 /**
  * Module to generate commerce and product related entries.
